generalized network queryer tests
